requirements.txt
opencv-python>=4.8.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
moviepy>=1.0.3
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
scipy>=1.11.0